smart-fish-farming-RELEASE/requirements.txt
APScheduler==3.9.1
backports.zoneinfo==0.2.1
certifi==2022.9.24
charset-normalizer==2.1.1
idna==3.4
kafka-python==2.0.2
pytz==2022.4
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
six==1.16.0
tzdata==2022.4
tzlocal==4.2
urllib3==1.26.12
